Fixed Animation Insights traced data not appearing until PIE is stopped when enabling tracing while PIE is in progress
Logic to build anim tracks originally skipped if there were no event tracks present for an object (e.g. a BeginPlay for a component). Now we go directly to the animation provider to ask if data is present.

Overall much less tracks are created, meaning we are less likely to hit the 1000-track limit and stop visualizing data.

#jira UE-100576 - Animation Insights stops logging after first Run-Stop in PIE.
#rb Jurre.deBaare

#ROBOMERGE-SOURCE: CL 14490324 in //UE4/Release-4.26/...
#ROBOMERGE-BOT: RELEASE (Release-4.26 -> Release-Engine-Staging) (v748-14444958)

[CL 14490336 by thomas sarkanen in Release-Engine-Staging branch]