Fixed invalid RTV transitions when EResourceTransitionFlags::MaintainCompression is specified.
The SkipFastClearEliminateState constant contained the SRV state bits, and RHIEndTransitions() was blindly adding it in when the MaintainCompression flag was present. This led to and invalid destination state when transitioning to RTV. Changed RHIEndTransitions() to only add the flags when transitioning to a read-only state, since that's when decompression can happen anyway. Also changed the constant to only contain compression-specific flags (when the platform has them), so the code which uses it can add the actual destination state, instead of assuming it's SRV.

#jira none
#rnx
#rb Eric.Mcdaniel, Kenzo.Terelst

#ROBOMERGE-SOURCE: CL 14500253 in //UE4/Release-4.26/... via CL 14500262
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v748-14444958)

[CL 14500266 by mihnea balta in Main branch]