Exp Height Fog depth bound optim
Use depth bounds to early kill pixel that are before the FogStartDistance. We can use that optimization when we are not using the VolumetricFog, which is the case on base ps4 + XBoxOne base and XboxOneS.

Grain in average 0.37ms on ps4 and 0.34ms on XB1. Increase dynamic resolution by 1% on Xb1S.

#rb [at]sebastien.hillaire
#jira none

#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: serge.bernier
#ROBOMERGE-SOURCE: CL 14494086 via CL 14494158 via CL 14494928 via CL 14495347 via CL 14495395
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v748-14444958)

[CL 14509768 by serge bernier in Dev-Tools-Staging branch]