Fix adaptive sampling in path tracer
There were a number of issues that resulted in NaNs creeping up to the top level of the hierarchy. This in turn led to the adaptive codepath being consistently disabled. I also fixed the incremental calculation of variance (we want to measure variance of the mean) although it doesn't seem to be possible to compute an exact variance when we have more than one sample per pixel from a single pass (in other words, when adaptive has started focusing on some pixels).

I switched the gaussian blur to a simple max dilation to be a bit more robust to outlier pixels.

Despite the adaptive sampler working fairly well on high error regions, I found it generally slows down rendering significantly, so I opted to disable it by default.

#rb Patrick.Kelly

[CL 14629658 by chris kulla in ue5-main branch]