Fix missing hair shadow when using lights which don't casting shadow.
The error relates to contact shadow changes done in CL 14102304 / 14125913. Exponential attenuation shouldn't be applied on hair, as the hit distance is used in place of the estimated thickness (complementary of the hit distance from the light blocker). This creates light leaks. This CL reverts to the old behavior for hair pixels, i.e., light was blocked on the first hit.

This CL also simplifies the application of contact shadow on transmittance shadow, as now hair shading is unified, and special code is no longer needed.

#rb none
#jira UE-101685, UE-102278
[FYI] andrew.lauritzen

#ROBOMERGE-SOURCE: CL 14662087 in //UE4/Release-4.26/...
#ROBOMERGE-BOT: RELEASE (Release-4.26 -> Release-Engine-Staging) (v756-14651656)

[CL 14662090 by charles derousiers in Release-Engine-Staging branch]