SharedBuffer: Extended shared buffer with an optional outer buffer and optional buffer owner
MakeView now has overloads that accept an outer shared buffer that it will hold a reference to and inherit the read-only state from. The view must be contained by the outer shared buffer.

TakeOwnership now has overloads that accept an instance of a type derived from FSharedBuffer::FBufferOwner that will be moved and stored to manage the lifetime of the buffer when a delete function is insufficient.

Functions that create buffers now return pointers instead of references, for more optimal reference counting. ToSharedRef can be called on the return value to convert it to a reference in cases where that is useful.

#jira none
#rb Zousar.Shaker
#rnx

[CL 14680479 by Devin Doucette in ue5-main branch]