Fix regression in directional light ray-traced shadows.
TraceDistance was introduced to clip the directional light shadow extent. This was done as a significant performance optimization for subsurface rays which cast from the point-of-view of the light source, as the original directional light shadowing distance is effectively unbounded. Unfortunately, applying TraceDistance has the adverse effect of changing default shadowing behavior in large scenes.

This change only applies TraceDistance to clip the distance of a subsurface ray. This prevents the shadowing regression in large scenes but keeps the performance optimization for subsurface rays.

#rb juan.canada
#jira ue-102522
#lockdown marcus.wassmer

#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: patrick.kelly
#ROBOMERGE-SOURCE: CL 14652962 in //UE4/Release-4.26/... via CL 14653021 via CL 14653039
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v760-14685797)

[CL 14706371 by patrick kelly in Dev-Tools-Staging branch]