Remote Session protocol now has a disconnect function/event that can be used to gracefully close a connection. This allows hosts or clients to immediately return to a waiting state instead of having to timeout first.
Updated code that uses remote session to provide a reason that can be displayed on the host/client when a disconnect occurs. The RemoteSessionRole is now responsible for determining when pings are necessary and if a timeout should occur. The lower level connections simply provide information about the last time certain events occurred.

Receiving 'ping' packets is now only necessary for keeping a connection alive if no other packets are being received. Connections will also respond with a 'pong' that allows latency to be measured.

#jira UE-103108, UE-103100
#rb swarm
#review @david.hibbitts

[CL 14749637 by Andrew Grant in 4.26 branch]