Sequencer: Control Rig. Toolbar Start and Pose Library
ControlRigPose contains the Pose Data UStructs the Pose Asset UObject.
ControlRigPoseMirrorTable contains logic for doing mirroring of controls.
SControlRigBaseListWidget is the main Pose Dialog. It's made to handle all three types of assets (pose,animation,selection), but
now is just hosting poses. You'll see that I left in code (commented out mostly) if and when we get these other types in.
SControlRigControlViews holds the specific View for the selected Asset, again just Poses.
The Thumbnail renderer is for automatically creating thumbnails (done it's own scene like Sequencer).. We went back and forth between that and taking a viewport snapshot, settling on the viewport snapshot instead. So like some other stuff may get removed but easier to do later than delete and re-create.
The rest is mostly tooling to get this setup, needed a proper toolkit to host the toolbar, etc..
#jira UESEQ-749
#rb max.chen

[CL 14760815 by Mike Zyracki in ue5-main branch]