Niagara Performance Reports and Baseline Gathering.
Niagara Effect Types can now define a "Baseline" system which other systems of this type can be compared against for a rough estimate of their relative performance that will work in the Editor and across various platfroms.
The idea is to give a rough but actionalble guide to performance which the artists can use to identify systems that need optimization work without the help of the code team.
EffectTypes define a Controller object that will generate the systems and control them if they need user parameters settings etc. Basic type in C++ but custom controller can be done in BP.

Reporting:
We have 3 types of output for this data.
1. Stats UI displaying all systems, their current stats, their baseline stats and the ratio between these.
2. A minimal reporter that will listen for any poorly or badly performing systems and dump them to the screen.
3. On world transitions we dump a more verbose report with all the above to the log.

Generation:
In editor we spawn a new window with any required baseline's that need regenerating. We store the results in per object config so it need only be done once.
In game we spawn all baselines into the game world on first tick so that we can still gather data in playtests.
Baseline systems must be constructed to ensure they don't cull or scale down during generation etc.


Refactored ParticlePerfStats:
Now using listener model so that serveral things can be gathering stats for different purposes without any interference with each other. Existing features moved over to work using listener classes.
Split between GT and RT stats now more explicit as we can't just flush the RT when we need the results in all cases.

- Refactored simple engine stats to use a delegate rather than a raw function pointer and added an API for plugins and other external code to add new stats. Allows Niagara code to register these simple debug rendering stat commands.
- PreviewLODDistance now applied to renderers to fake distance for ribbon tesselation and other distance based renderer level scaling.


#rb Stu.Mckenna, Shaun.Kime

#ROBOMERGE-OWNER: simon.tovey
#ROBOMERGE-AUTHOR: simon.tovey
#ROBOMERGE-SOURCE: CL 14820112 via CL 14820154 via CL 14820156
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v763-14813663)

[CL 14820165 by simon tovey in Main branch]