Fixed Mirroring. Now use similar setup to Control Rig with just support for each axis insread of an aribtrary plane.
Added ability to Rename Controls in the Pose. Needed in case the rig changes name on it's controls, needed a fast way to select a whole bunch of poses and rename them.

Added set of Asset Actions for the right click context menu in the Asset Browser.

Now Support multiple folders to filter on with new Add Folder To View context menu item. This will also add the folder to the project settings so it shows up on in the next session also.

Added ability to Update a Pose, which basically does a save over the new pose. Will popup a dialog so a user can confirm this destructive behavior.

Fixed issue with the thumbnail not showing up in the Pose View, main issue that it seems we had to use the level editor thumbnail pool.

Made Pose Name in Pose View text box and not editable for now.
#jira UE-103293
#jira UE-103385
#jira UE-103366

#rb max.chen

[CL 14899797 by Mike Zyracki in ue5-main branch]