Control Rig: Fixed "Deleting all imported bones is not shown correctly the rig hierarchy tree view and leads to crash when click".
This happened because Ctrl+A selects the last placeholder item in addition to other regular rig elements. In the old logic, we loop through all selected items and when it hits the placeholder item it triggers an early return. As a result, refresh tree view is not getting called if this placeholder item is included in the selection.

#jira UE-104350

#rb helge.mathee

[CL 14928513 by jack cai in ue5-main branch]