Fixed pointer invalidation when new thread events were received.
Events marked as important are associated internally to a pseudo thread with id 0 [A]. New-thread events are considered important and thus appear as if they were traced on thread 0. They will add thread N which will modify the Infos array. If someone was holding a A& = Infos[0] when Infos[N] was added then A can become stale. To combat this edge-case we'll thread 0 as a special case so it doesn't invalidate itself with its own events.

#rb arciel.rekman (indirectly)
#rnx

[CL 14989274 by Martin Ridgers in ue5-main branch]