Lumen: Optimized hardware, ray-traced traversal.
Support for deferred material evaluation in Lumen:
Lumen now invokes the sorted-deferred material evaluation pipeline. It is currently missing the platform-specific BookmarkBuffer optimization introduced by Yuriy.

Support for a minimal lumen payload, with custom ray-traversal specialization on certain platforms:
A custom minimal payload is created to optimize the Lumen-specific case of ray traversal with surface-cache lighting. In these instances, only the hit t-value and instance-id should be required in the payload. Optional WorldNormal support should be added in the future, as we are currently relying on the SDF normal.

Support for Ray and TraceHitDistance rebasing:
Due to limitations on certain platforms, the ray and hit-distance is rebased such that the ray always begins with t-min=0. My understanding is that this is a reported SDK bug that will be adressed in the future, such that the rebasing can be removed.

Performance:
1) Deferred material evaluation PC timings show a 1.65X speedup for LightingMode=1 and 1.96X speedup for LightingMode=2.
2) Minimal payload PC timings show a 1.41X speedup for LightingMode=0.

Future work:
1) Introduce the bookmark buffer to accelerate secondary ray-traversal in sorted-deferred material execution.
2) Support for these accelerations in the hardware-accelerated Screen Probe Gather code path.

#rb daniel.wright, krzysztof@narkowicz

[CL 15013384 by Patrick Kelly in ue5-main branch]