#Jira UE-103644: Material slot issue on importing CATPart and retessellate
The probleme is due to the association of the material and the color to a face was done in a second time according to an index. This index was initiated in GetBodyTessellation. In the bug case, the repair split the body into set of bodies. From the second body, the association of material associated the material of the ith face of the sub body to the ith face.
To fix the bug, the association of material associated the material to the last procced face.
(the modification of function header will be done in 4.27)

#rb jeanluc.corenthin
#lockdown simon.tourangeau

#ROBOMERGE-SOURCE: CL 15039787 in //UE4/Release-4.26/... via CL 15039793 via CL 15039797
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v767-15032292)

[CL 15039804 by david lesage in Main branch]