#jira UE-106213
The probleme is that SceneGraphArchive file name is based only on cad file description.
So it is the same with "Stitching Sew" and "Stitching None"
if the user import the same file 3 times with "Stitching None", then "Stitching Sew" and "Stitching None"
In the Cache, the first import generate the geometry and scene graphe cache for "Stitching None"
The second import generate the geometry cache for "Stitching Sew" and replace the scene graphe cache (as scene graphe cache is the same for all stitching technique.
The last import find the the cache for "Stitching None" but the scene graphe cache is for "Stitching Sew". So it build the scene graphe of Stitching Sew with geometry of "Stitching None". As there is only one mesh, a message of missing static mesh appear.

To Fix it:
- SceneGraphArchive file name is based on the cad file description and stitching technique
- MeshArchive file name is based on the SceneGraphArchive and meshing technique

To do in 4.27:
- CT file name must be saved in the class to avoid computing it twice

#rb jeanluc.corenthin
#lockdown simon.tourangeau

#ROBOMERGE-SOURCE: CL 15067177 in //UE4/Release-4.26/... via CL 15067197 via CL 15067202
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v767-15032292)

[CL 15067208 by david lesage in Main branch]