Prototype Wrapped Object Pointers (ObjectHandle/ObjectPtr)
Change cast of wrapped pointer to return raw pointer instead of wrapped pointer.  This is because the cast may return an IInterface instead of a UObject, and TObjectPtr<IInterface> is invalid and can result in some code (eg: access tracking) interpreting the pointer as a UObject, when that is not valid.  Could be modified in the future to return a wrapped pointer if the type being cast to is a UObject, and a raw pointer if the type being cast to is an IInterface.

#rb devin.doucette

[CL 15130801 by Zousar Shaker in ue5-main branch]