Prototype Wrapped Object Pointers (ObjectHandle/ObjectPtr)
Change the FObjectPtrProperty reflection type to derive from FObjectProperty type (instead of TFObjectPropertyBase) to avoid incompatibility with code that uses the reflection system to operate on specifically FObjectProperty fields.  Also removed the override of the GetID virtual method to allow ObjectPtr properties to have the same identifying name as Object properties.  This ensures there's no serialization incompatibility when moving between Object<->ObjectPtr or ObjectPtr->(Object compatible type) in the serialization and reflection code.

#jira UE-106329
#rb devin.doucette

[CL 15132777 by Zousar Shaker in ue5-main branch]