Modification of the skeletal mesh section chunking algorithm. The previous version was develop for character using a skeleton containing 100 to 200 bones. The algorithm separate the mesh in triangle island patch, a serie of connected triangle where the connection shared all vertices properties (position, tangentes, normals, vertex color, UVs in channel 0). Then after creating the mesh patch we merge together the patch until we reach the MBC (maximum bone per chunk) value. The result of this algo if all the mesh patch do not use more bones then the MBC value will create chunk that do not add vertices to the end result.
Why would we want to change this? Because if we have a skeletal mesh with only one mesh patch but using several hundreds of bones, we will have to create multiple
chunks for one mesh patch. The previous algorithm in that case will create patch randomly and the result can be catastrophic on the number of chunk it create and
there will be one draw call per chunk since each chunk create a skeletalmesh section. This change try to optimize the number of chunk generate by a single mesh patch.

The optimization is very simple. When we iterate a triangle in a mesh patch, we try to fit the triangle in all previous existing chunks (in the mesh patch) that already
use bone up to the MBC value. If the bones use by the triangle are all contain in such a previous chunk, we add the triangle into this chunk. Simply doing this can reduce
a lot the number of section for a single mesh patch. I saw on some test model above 50% reduction of the chunk number generated by a mesh patch.

For character that do not have mesh patch that use more bone then the MBC value, the algorithm give the same result.

On a performance side I see a small increment of the chunking time, but a reduce number of vertices and sections for mesh patch using more then the MBC value
MBC value of 75

Character body, no mesh patch over the MBC value:
*** BEFORE Chunk Time:ms  Triangles:23668  Vertices:14871  Bones:104  Sections: 2
*** NOW Chunk Time:29ms  Triangles:23668  Vertices:14871  Bones:104  Sections: 2
Result: Same result here since there is no mesh patch over MBC value

Plane, 23 mesh patch 13 under MBC and 10 over the MBC
*** BEFORE Chunk Time:40ms  Triangles:20000  Vertices:17360  Bones:1681 Sections: 72
*** NOW Chunk Time:44ms  Triangles:20000  Vertices:12267  Bones:1681  Sections: 51
Result: There is a several reduction of the number of vertices generate by the chunking (5K less vertices). The number of generated sections by the mesh patch over MBC reduce from (72-13)59 to (51 - 13)38

Head Mesh with only one mesh patch over the MBC value
*** BEFORE Chunk Time:230ms  Triangles:64588  Vertices:50565  Bones:892 Sections: 65
*** NOW Chunk Time:275ms  Triangles:64588  Vertices:37524  Bones:892  Sections: 35
Result: There is a several reduction of the number of vertices generate by the chunking (13K less vertices). The number of generated sections by the single mesh patch reduce from 65 to 35

By default the new modification is turn off and you need to set a console variable to turn on the modification.
You can add to the Engine.ini the following to turn on the feature.

[SystemSettings]
SkeletalMesh.UseExperimentalChunking=1

In the next major engine version this modification will be on by default.

#jira DH-2889
#rb richard.talbotwatkin
#lockdown cristina.riveron

#ROBOMERGE-SOURCE: CL 15144571 in //UE4/Release-4.26/... via CL 15144595 via CL 15144706
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v771-15082668)

[CL 15145211 by alexis matte in Main branch]