Prototype Wrapped Object Pointers (ObjectHandle/ObjectPtr)
Avoid an assert experienced when wrapped object pointers are used in named values via the StructuredArchive's FArchiveFromStructuredArchive adapter.  This change ensures that the wrapped pointers are serialized directly instead of through the adapter.  The groundwork for avoiding the adapter was already there, but what was missing was the operator necessary to convert from the templated TObjectPtr class to the private non-templated FObjectPtr base class used by serialization.  I suspect this same issue may also be happening for other types like TLazyObjectPtr which don't implement the necessary operator to convert to their private base type used in serialization.

#jira UE-106345
#rb devin.doucette
#fyi johan.torp

[CL 15162153 by Zousar Shaker in ue5-main branch]