DX12 workaround:  Force timestamp queries to graphics commmand list
We are currently seeing busted GPU stats due to timestamp queries executed on compute commandlists failing to get updated timestamps in the results buffer.  As a workaround, I am forcing the timestamp query resolves to a graphics commandlist with this change.  This will be revisited and is tracked with UE-107558.

#rb David.Harvey

#ROBOMERGE-SOURCE: CL 15274478 via CL 15274668 via CL 15274684
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v771-15082668)

[CL 15274696 by eric mcdaniel in Main branch]