Move SetCachedMaxDrawDistance to happen before Super::PostEditChangeProperty().
AActor::RerunConstructionScripts is destroying a UPrimitiveComponent which is currently being run through PostEditChangeProperty().  When UPrimitiveComponent ::PostEditChangeProperty() goes to cache off values based on the changed property they are cached on a TRASHED component that isn't in the world anymore.  Since this is known behavior you must cache values off on your component _before_ calling Super::PostEditChangeProperty().

This issue only happens when editing Blueprint constructed components that are attached to the actor Blueprint.

I'm not entirely happy with this fix because it only fixes this single property and doesn't account for potential other bugs.  I also explictly didn't move the Navigation and StreamingManager notifications to happen before Super::PostEditChange() because I don't know the side effects from moving that code to happen before.

#jira FROST-929
#review-15291590 @arciel.rekman, @mihnea.balta

[CL 15295953 by Jason Nadro in ue5-main branch]