#jira UE-107755
Editor layout system: If there is a version bump in the layout system, it used to show a message dialog always. Behavior modified.
If the load request comes from the "Load" UI button, it keep displaying a message dialog to warn the user (and a UELOG(Warning)). If it's from the editor startup, it will only show the UE_LOG(Warning).

Also modified the default layout with the new version.

#rb none

[CL 15307798 by Gines Hidalgo in ue5-main branch]